cli: hide command description for misuses
When a user misuses the CLI program, usage instructions are spitted to
them. This commit simply strips the command's description, not the usage
itself, since the user probably already knows what the command does,
just couldn't correctly input the correct arguments or subcommands.

Description is preserved if the user explicitly asks for help either
using `-h` or `-help` options.